WKS-928 - Expose commands
