+++ main.go
@@ -2,33 +2,9 @@ package main
 
 import (
 	"github.com/jfrog/jfrog-cli-core/v2/plugins"
-	"github.com/jfrog/jfrog-cli-core/v2/plugins/components"
-
-	"github.com/jfrog/workers-cli/commands"
+	"github.com/jfrog/workers-cli/cli"
 )
 
 func main() {
-	plugins.PluginMain(getApp())
-}
-
-func getApp() components.App {
-	app := components.App{}
-	app.Name = "worker"
-	app.Description = "Tools for managing workers"
-	app.Version = "v1.0.0"
-	app.Commands = getCommands()
-	return app
-}
-
-func getCommands() []components.Command {
-	return []components.Command{
-		commands.GetInitCommand(),
-		commands.GetDryRunCommand(),
-		commands.GetDeployCommand(),
-		commands.GetExecuteCommand(),
-		commands.GetRemoveCommand(),
-		commands.GetListCommand(),
-		commands.GetAddSecretCommand(),
-		commands.GetListEventsCommand(),
-	}
+	plugins.PluginMain(cli.GetApp())
 }
